Program to interfaces: core module
DEVSIX-424

+++ core/src/main/java/com/itextpdf/core/pdf/colorspace/PdfCieBasedCs.java
@@ -136,7 +136,7 @@ abstract public class PdfCieBasedCs extends PdfColorSpace<PdfArray> {
 
         @Override
         public int getNumOfComponents() {
-            return ((PdfArray) getPdfObject()).getAsStream(1).getAsInt(PdfName.Action.N).intValue();
+            return ((PdfArray) getPdfObject()).getAsStream(1).getAsInt(PdfName.Action.N);
         }
